Added VGG embedding
