+++ .idea/workspace.xml
@@ -5,7 +5,7 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="e5325b1d-e37b-4345-b080-21ad24c9aa46" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/scripts/mri/validate.py" beforeDir="false" afterPath="$PROJECT_DIR$/scripts/mri/validate.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/utils/embeddings.py" beforeDir="false" afterPath="$PROJECT_DIR$/utils/embeddings.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -231,7 +231,7 @@
       <workItem from="1666711407540" duration="2838000" />
       <workItem from="1668701794495" duration="2725000" />
       <workItem from="1668719799086" duration="791000" />
-      <workItem from="1684963294638" duration="13427000" />
+      <workItem from="1684963294638" duration="13533000" />
     </task>
     <servers />
   </component>
